==> Notebooks/01 - Gerar dataset da tabela Combustíveis a partir das planilhas.ipynb <==
import pandas as pd
import numpy as np
import os
import unicodedata


#Definição de variáveis
pasta_datasets='../Datasets/'
pasta_planilhas='../Planilhas/Combustivel/'

try:
    os.mkdir(pasta_datasets)
    print(f'"A pasta {pasta_datasets} foi criada.') 
except FileExistsError:
    print(f'"A pasta {pasta_datasets} já existe.') 
    


# Função utilizada para remover acentos e demais caracteres especiais de uma string
def ajustar_string(texto):
    texto_ajustado = unicodedata.normalize("NFD", texto)
    texto_ajustado = texto_ajustado.encode("ascii", "ignore")
    texto_ajustado = texto_ajustado.decode("utf-8")
    return texto_ajustado
    

# Cria uma lista com todos os arquivos contidos na pasta_planilhas, os quais serão processados e agrupados
planilhas = os.listdir(pasta_planilhas)

# Cria um dicionário para receber um dataframe por arquivo
d_df={}

# Defina as colunas de interesse
colunas = ['Regiao - Sigla',
           'Estado - Sigla',
           'Municipio',
           #'CNPJ da Revenda',
           'Produto',
           'Data da Coleta',
           'Valor de Venda',
           'Unidade de Medida',
           #'Bandeira'
          ]

# Define os tipos dos dados de cada coluna de interesse
tipos_dados_colunas = {'Regiao - Sigla':str,
                       'Estado - Sigla':str,
                       'Municipio':str,
                       #'CNPJ da Revenda':str,
                       'Produto':str,
                       'Data da Coleta':str,
                       'Valor de Venda':np.float64,
                       'Unidade de Medida':str
                       #'Bandeira':str
                      }

# Percorre todos os arquivos em 'pasta_planilhas'
for planilha in planilhas:
    
    df = pd.read_csv(pasta_planilhas+planilha,
                     sep = ';',
                     decimal=',',
                     low_memory=False, 
                     parse_dates=['Data da Coleta'],
                     dayfirst = True,encoding='utf-8'
                    )
    
    # Como constatou-se que alguns arquivos possuem acentos nos nomes das colunas, optou-se por utilizar a função
    # "ajustar_string" para normalizar estes nomes adequando-os ao padrão do dicionário 'colunas' acima
    for coluna_no_df in df.columns:
        df.rename(columns ={coluna_no_df:ajustar_string(coluna_no_df)}, inplace = True)
    
    # Agrega as colunas de interesse do dataframe lido no dicionário, utilizando como chave o nome do arquivo
    d_df[planilha] = df[colunas]
    
print('Foram inseridos ' + f'{len(d_df)} dataframes no dicionário d_df\n')
print(planilhas)

# Concatena os dataframes num único dataframe
df_combustiveis = pd.concat(d_df,ignore_index=True,sort=False)

display(df_combustiveis)
df_combustiveis.info()


# Simplifica o nome de algumas colunas
colunas_renomeadas = {'Data da Coleta': 'Data',
                      'Estado - Sigla' :'Estado',
                      'Regiao - Sigla' : 'Regiao',
                      'Valor de Venda' : 'ValorVenda',
                      'Unidade de Medida' : 'UnMed'
                     }

df_combustiveis.rename(columns=colunas_renomeadas, inplace = True)

print(df_combustiveis.columns)
df_combustiveis.info()

# Verifica se há algum valor nulo no dataframe 'df_combustiveis'

df_combustiveis.isnull().values.any()

# Verifica os tipos de combustíveis disponíveis na coluna Produto.
df_combustiveis['Produto'].unique()

# Como constatou a existência do produto 'ETANO', verificou-se se era um erro de tipo, e se 
# este dado deveria ser tratado
display(df_combustiveis[df_combustiveis.Produto == 'ETANO'])

# Neste estudo serão analisados apenas os combustíveis 'GASOLINA' e 'ETANOL'. 
# Não serão considerados assim a 'GASOLINA ADITIVADA', 'DIESEL' e suas variantes e também 'GNV'
# Como o dado do produto 'ETANO' é um erro, conforme acima, ele também será ignorado.
aceitar = ['GASOLINA','ETANOL']

df_combustiveis.drop(df_combustiveis[~df_combustiveis.Produto.isin(aceitar)].index, inplace=True)

display(df_combustiveis)

# Verifica se há mais de uma unidade de medida
df_combustiveis['UnMed'].unique()

# Como a unidade de medida é sempre a mesma, a coluna "UnMed" não será mais necessária
df_combustiveis.drop(columns='UnMed', inplace=True)

df_combustiveis.info()

# Verifica se há algum valor nulo no dataframe 'df_combustiveis'
df_combustiveis.isnull().values.any()

# Cria um novo índice com a coluna 'Data' 
df_combustiveis.set_index(['Data'], inplace=True)

display(df_combustiveis)

# Verifica a menor e a maior data no dataframe 'df_combustiveis'
menor_data = df_combustiveis.index.min()
maior_data = df_combustiveis.index.max()

print('Menor data no dataframe df_combustiveis: ' + str(menor_data))
print('Maior data no dataframe df_combustiveis: ' + str(maior_data))

# Exporta o dataset para um arquivo CSV
df_combustiveis.to_csv(pasta_datasets+'combustiveis.csv', sep = ';',index=True)
