==> Notebooks/02 - Efetuar download da planilha Dólar.ipynb <==
#pip install wget

import pandas as pd
import os
import wget


#Definição de variáveis
pasta_planilhas='../Planilhas/'
planilha_combustiveis = '../Datasets/combustiveis.csv'

# Função utilizada para remover um arquivo caso ele exista
# Foi preferível utilizar esta função ao invés de diretamente o os.remove() para efetuar o 
# tratamento de exceções e não interromper o processamento caso ocorram.

def remover_arquivo(url):
    try:
        os.remove(url)
        print('O arquivo ' + url + ' foi removido com sucesso.')
        return True
    except:
        #print('O arquivo ' + url + ' não pode ser removido.')
        return False

# Para efetuar o download da planilha Dólar, é necessário entrar como parâmetros, entre outros, o período desejado.
# Dessa forma, inicialmente é feita uma verificação no dataframe 'df_combustiveis' para obter esses valores.

# Verifica se o arquivo 'combustiveis.csv' existe
planilha_combustiveis_existe = os.path.isfile(planilha_combustiveis)

if planilha_combustiveis_existe:
    print(f'O arquivo {planilha_combustiveis} existe.') 
else:
    print(f'O arquivo {planilha_combustiveis} não existe, execute o Notebook anterior.') 


# Efetua a leitura do arquivo 'combustiveis.csv' 
df_combustiveis =  pd.read_csv(planilha_combustiveis,sep = ';',low_memory=False, parse_dates=['Data'],dayfirst = True,encoding='utf-8')

# Cria um novo índice com a coluna 'Data' 
df_combustiveis.set_index(['Data'], inplace=True)

display(df_combustiveis)
df_combustiveis.info()

# Verifica a menor e a maior data no dataframe 'df_combustiveis'
menor_data = df_combustiveis.index.min()
maior_data = df_combustiveis.index.max()

print('Menor data no dataframe df_combustiveis: ' + str(menor_data))
print('Maior data no dataframe df_combustiveis: ' + str(maior_data))

# Função utilizada para converter datetime em string no formato d/m/YYYY (exemplo: '2005-01-03 00:00:00' em '3/1/2005')
def ajustar_data_DMY(texto):
    texto_ajustado = texto.strftime('%#m/%#d/%Y')
    return texto_ajustado

# É efetuado o download da planilha Dolar a partir da plataforma Olinda, do BACEN, utilizando como parâmetro o período abrangido no dataframe df_combustiveis
# URL do site: https://olinda.bcb.gov.br/olinda/servico/PTAX/versao/v1/aplicacao#!/recursos/CotacaoDolarPeriodo#eyJmb3JtdWxhcmlvIjp7IiRmb3JtYXQiOiJqc29uIiwiJHRvcCI6MTAwfX0=
# URL de referência: https://olinda.bcb.gov.br/olinda/servico/PTAX/versao/v1/odata/CotacaoDolarPeriodo(dataInicial=@dataInicial,dataFinalCotacao=@dataFinalCotacao)?@dataInicial='1-1-2004'&@dataFinalCotacao='10-31-2021'&$top=10000&$orderby=dataHoraCotacao%20asc&$format=text/csv

url = 'https://olinda.bcb.gov.br/olinda/servico/PTAX/versao/v1/odata/CotacaoDolarPeriodo(dataInicial=@dataInicial,dataFinalCotacao=@dataFinalCotacao)?@dataInicial=' + '\'' + ajustar_data_DMY(menor_data) + '\''+'&@dataFinalCotacao=' + '\'' + ajustar_data_DMY(maior_data) + '\'' + '&$top=10000&$orderby=dataHoraCotacao%20asc&$format=text/csv'   

print(url)

try:
    # Inicialmente exclui o arquivo caso ele exista.
    remover_arquivo(pasta_planilhas +'planilha_dolar_bacen.csv')
    
    filename = wget.download(url,out=pasta_planilhas +'planilha_dolar_bacen.csv')
    print('O download da planilha \'planilha_dolar_bacen.csv\' foi efetuado com sucesso.')
except: 
    print('Não foi possível efetuar o download da planilha \'planilha_dolar_bacen.csv\'.')
