==> Notebooks/03 - Gerar o dataset da tabela Dólar a partir da planilha.ipynb <==
import pandas as pd
import numpy as np
import os

#Definição de variáveis
pasta_datasets='../Datasets/'
pasta_planilhas='../Planilhas/'
planilha_dolar_original = 'planilha_dolar_bacen.csv'

try:
    os.mkdir(pasta_datasets)
    print(f'"A pasta {pasta_datasets} foi criada.') 
except FileExistsError:
    print(f'"A pasta {pasta_datasets} já existe.') 

# Verifica se o arquivo 'planilha_dolar_original.csv' existe
planilha_dolar_existe = os.path.isfile(pasta_planilhas+planilha_dolar_original)

if planilha_dolar_existe:
    print(f'O arquivo {planilha_dolar_original} existe.') 
else:
    print(f'O arquivo {planilha_dolar_original} não existe, execute o Notebook anterior.') 


# Define os tipos dos dados das colunas de interesse

# Efetua a leitura do arquivo 'planilha_dolar_original.csv'  no dataframe 'df_dolar'
df_dolar = pd.read_csv(pasta_planilhas+planilha_dolar_original,sep = ',', decimal=',',low_memory=False, 
                       parse_dates=['dataHoraCotacao'],dayfirst = True,encoding='utf-8')

display(df_dolar)

df_dolar.info()


# Simplifica o nome da coluna dataHoraCotacao
colunas_renomeadas = {
                      'dataHoraCotacao' : 'Data'
                     }

df_dolar.rename(columns=colunas_renomeadas, inplace = True)

print(df_dolar.columns)
df_dolar.info()

# Verifica se há algum valor nulo no dataframe 'df_dolar'

df_dolar.isnull().values.any()

# Efetua a conversão do formato da coluna Data para o formato d/m/YYYY (exemplo: '2005-01-03 18:35:00.000' em '2005-01-03')
df_dolar['Data'] = pd.to_datetime(df_dolar['Data'].dt.strftime('%Y-%m-%d'), dayfirst = True)

display(df_dolar)

# Cria um novo índice com a coluna 'Data' 
df_dolar.set_index(['Data'], inplace=True)

display(df_dolar)

# Verifica a menor e a maior data no dataframe 'df_combustiveis'
menor_data = df_dolar.index.min()
maior_data = df_dolar.index.max()

print('Menor data no dataframe df_combustiveis: ' + str(menor_data))
print('Maior data no dataframe df_combustiveis: ' + str(maior_data))

# Exporta o dataset para um arquivo CSV
df_dolar.to_csv(pasta_datasets + 'dolar.csv', sep = ';',index=True)
