==> Notebooks/04 - Efetuar download da planilha IPCA.ipynb <==
#pip install wget

import pandas as pd
import os
import wget
import zipfile

#Definição de variáveis
pasta_planilha='../Planilhas/'

# Função utilizada para remover um arquivo caso ele exista
# Foi preferível utilizar esta função ao invés de diretamente o os.remove() para efetuar o 
# tratamento de exceções e não interromper o processamento caso ocorram.

def remover_arquivo(url):
    try:
        os.remove(url)
        print('O arquivo ' + url + ' foi removido com sucesso.')
        return True
    except:
        #print('O arquivo ' + url + ' não pode ser removido.')
        return False

# É efetuado o download da planilha IPCA a partir do FTP do IBGE.
# Essa planilha vem no arquivo 'ipca_SerieHist.zip', obtido no link abaixo, e é atualizada mensalmente com as 
# informações do IPCA referente ao mês imediatamente anterior.
# URL do site: https://ftp.ibge.gov.br/Precos_Indices_de_Precos_ao_Consumidor/IPCA/Serie_Historica/ipca_SerieHist.zip

url = 'https://ftp.ibge.gov.br/Precos_Indices_de_Precos_ao_Consumidor/IPCA/Serie_Historica/ipca_SerieHist.zip'   

print(url)

# Inicialmente exclui o arquivo caso ele exista.
remover_arquivo(pasta_planilha +'ipca_SerieHist.zip')

try:
    filename = wget.download(url,out=pasta_planilha +'ipca_SerieHist.zip')
    print('O download do arquivo \'ipca_SerieHist.zip\' foi efetuado com sucesso.')
except: 
    print('Não foi possível efetuar o download do arquivo \'ipca_SerieHist.zip\'.')


# Obtém o nome do arquivo a ser extraído de 'ipca_SerieHist.zip'
# Por padrão este nome varia de acordo com o ANO e o MES da última informação do IPCA
# Ex: 'ipca_202110SerieHist.xls'

arquivo_zip_ipca = zipfile.ZipFile(pasta_planilha+'ipca_SerieHist.zip', 'r')

try:
    # Retorna o nome da planilha XLS
    nome_arquivo_extraido = arquivo_zip_ipca.namelist()[0]

    # Se o arquivo a ser extraído já existe na pasta, é removido primeiro
    remover_arquivo(pasta_planilha+nome_arquivo_extraido)

    # É feita a extração no arquivo 'ipca_SerieHist.zip'
    arquivo_zip_ipca.extractall(pasta_planilha)

    # Se o arquivo 'planilha_ipca.xls' já existe, é removido primeiro
    remover_arquivo(pasta_planilha+'planilha_ipca.xls')

    # O arquivo extraído é renomeado para 'planilha_ipca.xls'
    os.rename(pasta_planilha+nome_arquivo_extraido,pasta_planilha+'planilha_ipca.xls')
    
    print('O arquivo \'planilha_ipca.xls\' foi extraído com sucesso.')
except:
    print('Não foi possível extrair o arquivo \'planilha_ipca.xls\'.')

# Fecha o arquivo ZIP para poder ser apagado
arquivo_zip_ipca.close()

# O arquivo ZIP é apagado
os.remove(pasta_planilha+'ipca_SerieHist.zip')
