==> Notebooks/06 - Efetuar o merge das tabelas Dólar e IPCA.ipynb <==
import pandas as pd
import os

# Define algumas variáveis
pasta_datasets = '../Datasets/'
arquivo_ipca = 'ipca_mensal.csv'
arquivo_dolar = 'dolar.csv'

# Efetua a leitura do arquivo 'ipca.csv' 

print(pasta_datasets+arquivo_ipca)
df_ipca =  pd.read_csv(pasta_datasets+arquivo_ipca,sep = ';',low_memory=False, parse_dates=['Data'],dayfirst = True,encoding='utf-8')

# Cria um novo índice com a coluna 'Data' 
df_ipca.set_index(['Data'], inplace=True)

display(df_ipca)
df_ipca.info()



# Efetua a leitura do arquivo 'dolar.csv' 

print(pasta_datasets+arquivo_dolar)
df_dolar =  pd.read_csv(pasta_datasets+arquivo_dolar,sep = ';',low_memory=False, parse_dates=['Data'],dayfirst = True,encoding='utf-8')

# Cria um novo índice com a coluna 'Data' 
df_dolar.set_index(['Data'], inplace=True)

display(df_dolar)
df_dolar.info()


# É efetuado um merge "outer" entre os dataframes df_dolar e df_ipca
# Cabe observar que o df_ipva possui dados apenas no dia 01 de cada mês

df_dolar_ipca = pd.merge(df_dolar,df_ipca,on="Data",how="outer")

display(df_dolar_ipca)
df_dolar_ipca.info()

#É feita a ordenação por Data, de forma crescente
df_dolar_ipca.sort_values(by='Data', ascending=True, inplace = True)

display(df_dolar_ipca)

# O outer join agrupou assim os dois dataframes, e observamos duas situações:
# 1 - quando a Data existe nos dois, as colunas estão todas preenchidas
# 2 - quando a Data existe apenas em uma das tabelas, há colunas com valores nulos, que deverão ser tratados.
# Assim, como o valor do IPCA é mensal, e vale para todos os dias do mês, então será necessário
# replicar este valor nas demais datas do mês

temp_indice_dez93 = -1
temp_ipca_mensal = -1
    
for idx in df_dolar_ipca.index:
    
    if(pd.notnull(df_dolar_ipca.loc[idx,'Indice_Dez93'])==False):
        df_dolar_ipca.loc[idx,'Indice_Dez93'] = temp_indice_dez93
        df_dolar_ipca.loc[idx,'Ipca_Mensal'] = temp_ipca_mensal
    else:
        temp_indice_dez93 = df_dolar_ipca.loc[idx,'Indice_Dez93']
        temp_ipca_mensal = df_dolar_ipca.loc[idx,'Ipca_Mensal']
        
display(df_dolar_ipca)


# Verifica se ainda há linhas com valores nulos
df_dolar_ipca.isnull().values.any()

df_dolar_ipca.info()

# Ainda há 82 linhas em que as colunas 'cotacaoCompra' e 'cotacaoVenda' permanecem em branco.
# Esses são os casos em que o 1º dia do mês não é dia útil, assim essas linhas serão desconsideradas

df_dolar_ipca.dropna(subset = ['cotacaoCompra'], inplace=True)

display(df_dolar_ipca)

# Verifica que esse dataframe contém apenas dias úteis (dayofweek de 0 a 4)

# 0 = Segunda-feira
# 1 = Terça-feira
# 2 = Quarta-feira
# 3 = Quinta-feira
# 4 = Sexta-feira
# 5 = Sábado
# 6 = Domingo

df_dolar_ipca.index.dayofweek.unique()

# Verifica se ainda há linhas com valores nulos
df_dolar_ipca.isnull().values.any()

df_dolar_ipca.info()

# Exporta o dataset para um arquivo CSV
df_dolar_ipca.to_csv(pasta_datasets + 'dolar_ipca.csv', sep = ';',index=True)
