==> Notebooks/07 - Efetuar o merge das tabelas Combustíveis com a tabela Dólar_IPCA.ipynb <==
import pandas as pd
import os
import matplotlib.pyplot as plt


# Define algumas variáveis

pasta_datasets = '../Datasets/'
planilha_dolar_ipca = 'dolar_ipca.csv'
planilha_combustiveis = 'combustiveis.csv'

# Efetua a leitura do arquivo 'combustiveis.csv' 
df_combustiveis =  pd.read_csv(pasta_datasets+planilha_combustiveis,sep = ';',low_memory=False, 
                               parse_dates=['Data'],dayfirst = True,encoding='utf-8')

# Cria um novo índice com a coluna 'Data' 
df_combustiveis.set_index(['Data'], inplace=True)

display(df_combustiveis)
df_combustiveis.info()

# A partir daqui os dados de venda dos municípios não serão utilizados diretamente, mas será feita a média
# dos valores por combustível no estado num mesmo dia.

# Como a idéia é montar uma tabela agregando numa mesma linha/dia os valores de venda do Etanol e da Gasolina, 
# para facilitar a análise, serão feitas duas tabelas e depois serão agrupadas

# Inicia montando um dataframe para Gasolina, com os valores agrupados inclusive por Estado

df_gasolina_estados = pd.DataFrame(df_combustiveis[df_combustiveis["Produto"] =='GASOLINA'].
                                   groupby(by=['Data','Regiao','Estado','Produto'])
                                   ['ValorVenda'].mean())

display(df_gasolina_estados)

# Renomeia a coluna de ValorVenda para ValorVendaGasolina, a fim de diferenciar quando fizer o merge

df_gasolina_estados.rename(columns= {'ValorVenda':'ValorVendaGasolina'}, inplace = True)

display(df_gasolina_estados)

df_gasolina_estados.info()

# A coluna 'Produto' não é mais necessária, e poderá causar problemas ao efetuar o merge com o dataframe do
# Etanol, desta forma ela será removida

# Como esta coluna é parte de um multi-index, inicialmente é necessário resetar o índice desta coluna

df_gasolina_estados = df_gasolina_estados.reset_index(level=['Produto'])

display(df_gasolina_estados)

# Agora é possível removê-la
df_gasolina_estados = df_gasolina_estados.drop('Produto',1)

display(df_gasolina_estados)

# Verifica que não existem valores null
df_gasolina_estados.info()

# Da mesma forma, monta o dataframe para Etanol, com os valores agrupados inclusive por Estado

df_etanol_estados = pd.DataFrame(df_combustiveis[df_combustiveis["Produto"] =='ETANOL'].
                                 groupby(by=['Data','Regiao','Estado','Produto'])
                                 ['ValorVenda'].mean())

display(df_etanol_estados)

# Renomeia a coluna de ValorVenda para ValorVendaEtanol, a fim de diferenciar quando fizer o merge

df_etanol_estados.rename(columns= {'ValorVenda':'ValorVendaEtanol'}, inplace = True)

display(df_etanol_estados)

df_etanol_estados.info()

# A coluna 'Produto' não é mais necessária, e poderá causar problemas ao efetuar o merge com o dataframe da
# Gasolina, desta forma ela será removida

# Como esta coluna é parte de um multi-index, inicialmente é necessário resetar o índice desta coluna

df_etanol_estados = df_etanol_estados.reset_index(level=['Produto'])

display(df_etanol_estados)

# Agora é possível removê-la
df_etanol_estados = df_etanol_estados.drop('Produto',1)

display(df_etanol_estados)

# Verifica que não existem valores null
df_etanol_estados.info()

# Importante observar que havia 146238 linhas anteriormente, sendo 74334 relativas a Gasolina e 71904 a Etanol.

# Com os dataframes prontos, de Gasolina e Etanol, será feito o merge deles
df_combustiveis_agrupados_estado = pd.merge(df_gasolina_estados,df_etanol_estados,
                                            on=["Data","Regiao","Estado"],how="outer")

display(df_combustiveis_agrupados_estado)


# Verifica a existência de valores null
df_combustiveis_agrupados_estado.info()

# Verifica que existem dias nos quais não houve coleta de dados em determinado estado, 
# resultando em valores nulos

qtd_linhas = df_combustiveis_agrupados_estado.shape[0]
qtd_valores_gasolina = sum(pd.notnull(df_combustiveis_agrupados_estado['ValorVendaGasolina']))
qtd_valores_etanol = sum(pd.notnull(df_combustiveis_agrupados_estado['ValorVendaEtanol']))
qtd_valores_nulos_gasolina = sum(pd.isnull(df_combustiveis_agrupados_estado['ValorVendaGasolina']))
valores_nulos_etanol = sum(pd.isnull(df_combustiveis_agrupados_estado['ValorVendaEtanol']))

print('Total de linhas: ' + str(qtd_linhas))
print('Quantidade de valores de gasolina: ' + str(qtd_valores_gasolina))
print('Quantidade de valores nulos de gasolina: ' + str(qtd_valores_nulos_gasolina))
print('Quantidade de valores de etanol: ' + str(qtd_valores_etanol))
print('Quantidade de valores nulos de etanol: ' + str(valores_nulos_etanol))

# Optou-se por remover as linhas com valor nulo de combustíveis

df_combustiveis_agrupados_estado.dropna(subset = ['ValorVendaGasolina'], inplace=True)
df_combustiveis_agrupados_estado.dropna(subset = ['ValorVendaEtanol'], inplace=True)

display(df_combustiveis_agrupados_estado)
df_combustiveis_agrupados_estado.info()

# Exporta o dataset para um arquivo CSV
#df_combustiveis_agrupados_estado.to_csv(pasta_datasets + 'df_combustiveis_agrupados_estado.csv', sep = ';',index=True)

# Efetua a leitura do arquivo 'dolar_ipca.csv'

print(pasta_datasets+planilha_dolar_ipca)
df_dolar_ipca =  pd.read_csv(pasta_datasets+planilha_dolar_ipca,sep = ';',low_memory=False, parse_dates=['Data'],dayfirst = True,encoding='utf-8')

# Cria um novo índice com a coluna 'Data' 
df_dolar_ipca.set_index(['Data'], inplace=True)

display(df_dolar_ipca)
df_dolar_ipca.info()



# Remove a coluna cotaCompra, referente ao dólar de compra, e que não será utilizada

df_dolar_ipca = df_dolar_ipca.drop(['cotacaoCompra'],axis=1)

# Renomeia a coluna "cotacaoVenda" para "CotacaoDolarVenda"

df_dolar_ipca = df_dolar_ipca.rename(columns={'cotacaoVenda':'CotacaoDolarVenda'})

display(df_combustiveis_agrupados_estado)
df_combustiveis_agrupados_estado.info()

# O dataframe df_combustiveis_agrupados_estado possui um multi-index, e antes de efetuar o merge com o 
# dataframe df_dolar_ipca será necessário resetar as colunas 'Região' e 'Estado' daquele, a fim de utilizar 
# apenas a coluna 'Data' como chave

df_combustiveis_agrupados_estado = df_combustiveis_agrupados_estado.reset_index(level=['Regiao','Estado'])

display(df_combustiveis_agrupados_estado)

df_combustiveis_agrupados_estado.info()

# Efetua o merge com o dataframe df_dolar_ipca

df_combustiveis_dolar_ipca_estado = pd.merge(df_combustiveis_agrupados_estado,
                                             df_dolar_ipca,on="Data",how="left")

display(df_combustiveis_dolar_ipca_estado)


# Verifica a existência de valores null, sendo que:
# - para as colunas derivadas do dataframe df_dolar_ipca, verifica se que há 1753 valores nulos,
# o que será verificado a seguir

df_combustiveis_dolar_ipca_estado.info()

# Como as colunas cotacaoCompra, cotacaoVenda, Indice_Dez93 e Ipca_Mensal possuem a mesma quantidade de 
# valores nulos, optou-se por utilizar a cotacaoVenda para verificar quais as Datas (index) que não 
# possuem dados

datas_valores_nulos_dolar = df_combustiveis_dolar_ipca_estado[df_combustiveis_dolar_ipca_estado['CotacaoDolarVenda'].isnull()].index
datas_valores_nulos_dolar_distintos = df_combustiveis_dolar_ipca_estado[df_combustiveis_dolar_ipca_estado['CotacaoDolarVenda'].isnull()].index.unique()

print('Quantidade de datas com valores nulos na coluna CotacaoDolarVenda: ' + str(datas_valores_nulos_dolar.size))
print('Quantidade de datas distintas com valores nulos na coluna CotacaoDolarVenda: ' + str(datas_valores_nulos_dolar_distintos.size))

# Verifica que o dataframe df_dolar_ipca contém apenas dias úteis (dayofweek de 0 a 4)

# 0 = Segunda-feira
# 1 = Terça-feira
# 2 = Quarta-feira
# 3 = Quinta-feira
# 4 = Sexta-feira
# 5 = Sábado
# 6 = Domingo

df_dolar_ipca.index.dayofweek.unique()

# Verifica que há dias em que houve leitura aos finais de semana (dayofweek 5 e 6), o que explicaria a 
# ausência de alguns dados na tabela, já que não existem finais de semana em df_dolar_ipca

datas_valores_nulos_dolar_distintos.dayofweek.unique()

# Faz uma lista com os índices correspondentes aos dias úteis da semana

dias_uteis = [0,1,2,3,4]

# Lista as datas distintas 
datas_valores_nulos_dolar_distintos

# Percorre a lista removendo as datas que não são finais de semana, de forma a gerar uma lista 
# de datas para análise

offset = 0

datas_uteis_valores_nulos_dolar_distintos = datas_valores_nulos_dolar_distintos

for data in datas_valores_nulos_dolar_distintos:
    if data.dayofweek not in dias_uteis:
        datas_uteis_valores_nulos_dolar_distintos = datas_uteis_valores_nulos_dolar_distintos.delete(offset)
    else:
        offset = offset + 1
    
print('Quantidade de dias úteis: ' + str(datas_uteis_valores_nulos_dolar_distintos.size))

# Confirma que a lista não tem final de semana

datas_uteis_valores_nulos_dolar_distintos.dayofweek.unique()

# Fazendo uma análise nestas datas, conclui-se que, apesar de não serem finais de semana, realmente 
# não foram dias úteis, mas feriados (Carnaval, Tiradentes, Independência, Natal, entre outros)

datas_uteis_valores_nulos_dolar_distintos

# Desta forma, realmente não deveriam existir no dataframe df_dolar_ipca
# Assim será feita a remoção destas linhas que não são dias uteis

# ANTES

display(df_combustiveis_dolar_ipca_estado)
df_combustiveis_dolar_ipca_estado.info()

# DEPOIS

df_combustiveis_dolar_ipca_estado.dropna(subset = ['CotacaoDolarVenda'], inplace=True)

display(df_combustiveis_dolar_ipca_estado)
df_combustiveis_dolar_ipca_estado.info()

# Exporta o dataset para um arquivo CSV
df_combustiveis_dolar_ipca_estado.to_csv(pasta_datasets + 'df_combustiveis_dolar_ipca_estado.csv', sep = ';',index=True)

# Com o intuito de efetuar a análise de Machine Learning do dataset, a princípio será feita
# uma comparação utilizando o dataset com agrupamentos a nível nacional

df_combustiveis_agrupados_nacional = pd.DataFrame(df_combustiveis_agrupados_estado.groupby('Data')
                                                  [['ValorVendaGasolina','ValorVendaEtanol']].mean())

df_combustiveis_agrupados_nacional


# Exporta o dataset para um arquivo CSV
#df_combustiveis_agrupados_nacional.to_csv(pasta_datasets + 'df_combustiveis_agrupados_nacional.csv', sep = ';',index=True)

# Efetua o merge com o dataframe df_dolar_ipca

df_combustiveis_dolar_ipca_nacional = pd.merge(df_combustiveis_agrupados_nacional,df_dolar_ipca,on="Data",how="left")

display(df_combustiveis_dolar_ipca_nacional)


# Verifica a existência de valores null, sendo que:
# - as células ValorVendaGasolina e ValorVendaEtanol nulas são dias em que não houve coleta 
# - já para as colunas derivadas do dataframe df_dolar_ipca, verifica se que há 123 valores nulos,
# o que será verificado a seguir

df_combustiveis_dolar_ipca_nacional.info()

# Esses são os mesmos dias da análise anterior, que não são dias úteis e portanto não devem existir 
# no dataframe df_combustiveis_agrupados_nacional

datas_valores_nulos_dolar_distintos_nacional = df_combustiveis_dolar_ipca_nacional[df_combustiveis_dolar_ipca_nacional['CotacaoDolarVenda'].isnull()].index.unique()

datas_valores_nulos_dolar_distintos_nacional

# Assim será feita a remoção destas linhas que não são dias uteis

# ANTES

display(df_combustiveis_dolar_ipca_nacional)
df_combustiveis_dolar_ipca_nacional.info()

# DEPOIS

df_combustiveis_dolar_ipca_nacional.dropna(subset = ['CotacaoDolarVenda'], inplace=True)

display(df_combustiveis_dolar_ipca_nacional)
df_combustiveis_dolar_ipca_nacional.info()

# Verificando os campos nulos, constata-se que há um dia em que não consta o dado em ValorVendaGasolina

# df_combustiveis_dolar_ipca_nacional[df_combustiveis_dolar_ipca_nacional['ValorVendaGasolina'].isna()]

# Optou-se assim por eliminar esta linha, para que o dataframe não tenha nenhum valor nulo

# df_combustiveis_dolar_ipca_nacional.dropna(subset = ['ValorVendaGasolina'], inplace=True)

# display(df_combustiveis_dolar_ipca_nacional)
# df_combustiveis_dolar_ipca_nacional.info()

# Exporta o dataset para um arquivo CSV
df_combustiveis_dolar_ipca_nacional.to_csv(pasta_datasets + 'df_combustiveis_dolar_ipca_nacional.csv', sep = ';',index=True)
